renamed file util.py to run_query.py
I feel like doing this as suggested by a fellow classmate would help understand the purpose of the code easier in that file and therefore improve readability
