+++ README.md
@@ -33,7 +33,7 @@ Components: vending_machine_id, name, location
 - Get all the products that is being sold in a certain vending machine
 - Get all the vending machines installed in a certain location
 
-### How to run
+## How to run
 `cd` into the `app` directory then run the follow command in your terminal: `python3 __init__.py`
 
 ### How to test
