+++ app/vending_machines.py
@@ -4,7 +4,7 @@ vending_bp = Blueprint('vending', __name__, url_prefix='/')
 
 
 def import_fun():
-    from util import run_sql_script
+    from run_query import run_sql_script
     return run_sql_script
